+++ mattermost-plugin/server/plugin_test.go
@@ -8,9 +8,7 @@ import (
 	"net/http"
 	"net/http/httptest"
 	"testing"
-	"time"
 
-	"github.com/golang/mock/gomock"
 	"github.com/mattermost/mattermost-server/v6/model"
 	"github.com/stretchr/testify/assert"
 )
@@ -116,67 +114,5 @@ func TestSetConfiguration(t *testing.T) {
 
 		assert.Equal(t, "true", config.FeatureFlags["hello_world"])
 		assert.Equal(t, "true", config.FeatureFlags["myTest"])
-	})
-}
-
-func TestRunDataRetention(t *testing.T) {
-	th := SetupTestHelperMockStore(t)
-	plugin := Plugin{}
-	plugin.server = th.Server
-
-	now := time.Now().UnixNano()
-
-	t.Run("test null license", func(t *testing.T) {
-		th.Store.EXPECT().GetLicense().Return(nil)
-		_, err := plugin.RunDataRetention(now, 10)
-		assert.NotNil(t, err)
-		assert.Equal(t, ErrInsufficientLicense, err)
-	})
-
-	t.Run("test invalid license", func(t *testing.T) {
-		falseValue := false
-
-		th.Store.EXPECT().GetLicense().Return(
-			&model.License{
-				Features: &model.Features{
-					DataRetention: &falseValue,
-				},
-			},
-		)
-		_, err := plugin.RunDataRetention(now, 10)
-		assert.NotNil(t, err)
-		assert.Equal(t, ErrInsufficientLicense, err)
-	})
-
-	t.Run("test valid license, invalid config", func(t *testing.T) {
-		trueValue := true
-		th.Store.EXPECT().GetLicense().Return(
-			&model.License{
-				Features: &model.Features{
-					DataRetention: &trueValue,
-				},
-			})
-
-		count, err := plugin.RunDataRetention(now, 10)
-		assert.Nil(t, err)
-		assert.Equal(t, int64(0), count)
-	})
-
-	t.Run("test valid license, valid config", func(t *testing.T) {
-		trueValue := true
-		th.Store.EXPECT().GetLicense().Return(
-			&model.License{
-				Features: &model.Features{
-					DataRetention: &trueValue,
-				},
-			})
-
-		th.Store.EXPECT().RunDataRetention(gomock.Any(), int64(10)).Return(int64(100), nil)
-		plugin.server.Config().EnableDataRetention = true
-
-		count, err := plugin.RunDataRetention(now, 10)
-
-		assert.Nil(t, err)
-		assert.Equal(t, int64(100), count)
 	})
 }
